Fixing asyncio queue creation and usage by decoupling app, queues, and uvicorn config.
